fix: fix missing type definitions caused by broken `index.d.ts`
