+++ .changeset/big-apes-cover.md
@@ -1,0 +1,5 @@
+---
+"@near-js/biometric-ed25519": patch
+---
+
+Fix missing type definitions in editors caused by broken `index.d.ts`
